# swagger.yaml
openapi: 3.0.0
info:
  title: PC-Hardware-API
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:5000/api/v1
    variables: {}
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create Order
      operationId: CreateOrder
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CreateOrderRequest'
                - example:
                    products:
                      - product: 66aba47c6c10f2b0ca25d955
                        quantity: 1
                      - product: 66aba4726c10f2b0ca25d94f
                        quantity: 1
            example:
              products:
                - product: 66aba47c6c10f2b0ca25d955
                  quantity: 1
                - product: 66aba4726c10f2b0ca25d94f
                  quantity: 1
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - Orders
      summary: Get All Orders
      operationId: GetAllOrders
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /orders/66aba00fa5bce0052839a5a9:
    get:
      tags:
        - Orders
      summary: Get Single Order
      operationId: GetSingleOrder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    patch:
      tags:
        - Orders
      summary: Update Order Status
      operationId: UpdateOrderStatus
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UpdateOrderStatusRequest'
                - example:
                    status: paid
            example:
              status: paid
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateOrderRequest:
      title: CreateOrderRequest
      required:
        - products
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
          description: ''
      example:
        products:
          - product: 66aba47c6c10f2b0ca25d955
            quantity: 1
          - product: 66aba4726c10f2b0ca25d94f
            quantity: 1
    Product:
      title: Product
      required:
        - product
        - quantity
      type: object
      properties:
        product:
          type: string
        quantity:
          type: integer
          format: int32
      example:
        product: 66aba47c6c10f2b0ca25d955
        quantity: 1
    UpdateOrderStatusRequest:
      title: UpdateOrderStatusRequest
      required:
        - status
      type: object
      properties:
        status:
          type: string
      example:
        status: paid
...
security:
  - {}
...